improving plugin config decoder files disposition
