+++ core/src/drivers/plugins/python/shared/plugin_config_decode/__init__.py
@@ -1,0 +1,16 @@
+"""
+Plugin configuration decoding package
+"""
+
+from .plugin_config_contact import PluginConfigContract, PluginConfigError
+from .modbus_master_config_model import ModbusTcpConfig, ModbusIoPointConfig
+from .plugin_config_decoder import PluginInstanceConfig, PluginConfigDecoder
+
+__all__ = [
+    'PluginConfigContract',
+    'PluginConfigError',
+    'PluginInstanceConfig', 
+    'PluginConfigDecoder',
+    'ModbusIoPointConfig',
+    'ModbusTcpConfig'
+]
